rename components dir to ui for consistency
