+++ src/feat/Projects/index.ts
@@ -1,5 +1,5 @@
 export * from './ProjectViewer';
 export * from './ProjectProvider';
 export * from './hooks/useProjectViewer';
-export * from './components/ProjectItem';
+export * from './ui/ProjectItem';
 export * from './types';
